requirements.txt
Django==1.11
django-rest-framework #3.7.3
django-role-permissions
django-bootstrap-form
django-bootstrap3
django-rest-swagger

psycopg2
pulp